Data/requirements.txt

build==0.7.0
packaging==21.0
python-Levenshtein==0.12.2
regex==2022.1.18
SPARQLWrapper==1.8.5
torchtext==0.12.0
tqdm==4.62.3